convert columnselect dialog
